Some bug fixes related to handling the lists.

+++ Serie2_SO_1314V_Peter/UThread/UThreadInternal.h
@@ -60,8 +60,8 @@ typedef struct _UTHREAD {
 	UT_ARGUMENT      Argument; 
 	PUCHAR           Stack;
 	HANDLE		     WaitingThread;
-	DWORD			 InitialTime;
-	DWORD			 TimeToWait;
+	LONG			 InitialTime;
+	LONG			 TimeToWait;
 } UTHREAD, *PUTHREAD;
 
 //
